# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/btc/getrawblock": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "get raw block with blockhash",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ordx.btc"
                ],
                "summary": "get raw block with blockhash",
                "parameters": [
                    {
                        "type": "string",
                        "description": "blockHash",
                        "name": "blockHash",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful response",
                        "schema": {
                            "$ref": "#/definitions/bitcoind.RawBlockResp"
                        }
                    },
                    "401": {
                        "description": "Invalid API Key"
                    }
                }
            }
        },
        "/btc/tx": {
            "post": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "send Raw Transaction",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ordx.btc"
                ],
                "summary": "send Raw Transaction",
                "parameters": [
                    {
                        "description": "Signed transaction hex",
                        "name": "signedTxHex",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "Reject transactions whose fee rate is higher than the specified value, expressed in BTC/kB.default:",
                        "name": "maxfeerate",
                        "in": "body",
                        "schema": {
                            "type": "number"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful response",
                        "schema": {
                            "$ref": "#/definitions/bitcoind.SendRawTxResp"
                        }
                    },
                    "401": {
                        "description": "Invalid API Key"
                    }
                }
            }
        },
...
        "/inscription/address/{address}": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Get a list of Inscription for a specific address",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "index.inscription"
                ],
                "summary": "Get a list of Inscription for a specific address",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Address",
                        "name": "address",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Start index for pagination",
                        "name": "start",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Limit for pagination",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful response",
                        "schema": {
                            "$ref": "#/definitions/index.OrdInscriptionListResp"
                        }
                    },
                    "401": {
                        "description": "Invalid API Key"
                    }
                }
            }
        },
        "/inscription/genesesaddress/{address}": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Get a list of Inscription for a specific geneses address",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "index.inscription"
                ],
                "summary": "Get a list of Inscription for a specific geneses address",
                "parameters": [
                    {
                        "type": "string",
                        "description": "geneses address",
                        "name": "address",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Start index for pagination",
                        "name": "start",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Limit for pagination",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful response",
                        "schema": {
                            "$ref": "#/definitions/index.OrdInscriptionListResp"
                        }
                    },
                    "401": {
                        "description": "Invalid API Key"
                    }
                }
            }
        },
        "/inscription/id/{id}": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Get inscription with inscription id",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "index.inscription"
                ],
                "summary": "Get inscription with inscription id",
                "parameters": [
                    {
                        "type": "string",
                        "description": "inscription id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful response",
                        "schema": {
                            "$ref": "#/definitions/index.OrdInscriptionResp"
                        }
                    },
                    "401": {
                        "description": "Invalid API Key"
                    }
                }
            }
        },
        "/inscription/number/{number}": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Get inscription with inscription number",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "index.inscription"
                ],
                "summary": "Get inscription with inscription number",
                "parameters": [
                    {
                        "type": "string",
                        "description": "inscription Number",
                        "name": "number",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful response",
                        "schema": {
                            "$ref": "#/definitions/index.OrdInscriptionResp"
                        }
                    },
                    "401": {
                        "description": "Invalid API Key"
                    }
                }
            }
        },
        "/inscription/sat/{sat}": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Get a list of Inscription with sat",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "index.inscription"
                ],
                "summary": "Get a list of Inscription with sat",
                "parameters": [
                    {
                        "type": "string",
                        "description": "sat",
                        "name": "sat",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Start index for pagination",
                        "name": "start",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Limit for pagination",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful response",
                        "schema": {
                            "$ref": "#/definitions/index.OrdInscriptionListResp"
                        }
                    },
                    "401": {
                        "description": "Invalid API Key"
                    }
                }
            }
        },
...
        "/ordstatus": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "index status",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "index"
                ],
                "summary": "index status",
                "responses": {
                    "200": {
                        "description": "Successful response",
                        "schema": {
                            "$ref": "#/definitions/index.IndexStatusResp"
                        }
                    },
                    "401": {
                        "description": "Invalid API Key"
                    }
                }
            }
        }
    },
    "definitions": {
        "bitcoind.RawBlockResp": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 0
                },
                "data": {
                    "type": "string",
                    "example": ""
                },
                "msg": {
                    "type": "string",
                    "example": "ok"
                }
            }
        },
        "bitcoind.SendRawTxResp": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 0
                },
                "data": {
                    "type": "string",
                    "example": "ae74538baa914f3799081ba78429d5d84f36a0127438e9f721dff584ac17b346"
                },
                "msg": {
                    "type": "string",
                    "example": "ok"
                }
            }
        },
...
        "index.IndexStatusResp": {
            "type": "object",
            "properties": {
                "addressCount": {
                    "type": "integer"
                },
                "blessedInscriptions": {
                    "type": "integer"
                },
                "cursedInscriptions": {
                    "type": "integer"
                },
                "dbVersion": {
                    "type": "string"
                },
                "genesesAddressCount": {
                    "type": "integer"
                },
                "indexVersion": {
                    "type": "string"
                },
                "syncInscriptionHeight": {
                    "type": "integer"
                },
                "syncTransferInscriptionHeight": {
                    "type": "integer"
                }
            }
        },
        "index.OrdInscriptionAsset": {
            "type": "object",
            "properties": {
                "inscription": {
                    "$ref": "#/definitions/pb.Inscription"
                },
                "utxo": {
                    "type": "string"
                }
            }
        },
        "index.OrdInscriptionListData": {
            "type": "object",
            "properties": {
                "detail": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/index.OrdInscriptionAsset"
                    }
                },
                "start": {
                    "type": "integer",
                    "example": 0
                },
                "total": {
                    "type": "integer",
                    "example": 9992
                }
            }
        },
        "index.OrdInscriptionListResp": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 0
                },
                "data": {
                    "$ref": "#/definitions/index.OrdInscriptionListData"
                },
                "msg": {
                    "type": "string",
                    "example": "ok"
                }
            }
        },
        "index.OrdInscriptionResp": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 0
                },
                "data": {
                    "$ref": "#/definitions/index.OrdInscriptionAsset"
                },
                "msg": {
                    "type": "string",
                    "example": "ok"
                }
            }
        },
...
        "pb.Inscription": {
            "type": "object",
            "properties": {
                "OrdSat": {
                    "type": "integer"
                },
                "address": {
                    "type": "string"
                },
                "children": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "contentLength": {
                    "type": "integer"
                },
                "contentType": {
                    "type": "string"
                },
                "fee": {
                    "type": "integer"
                },
                "genesesAddress": {
                    "type": "string"
                },
                "height": {
                    "type": "integer"
                },
                "id": {
                    "type": "string"
                },
                "next": {
                    "type": "string"
                },
                "number": {
                    "type": "integer"
                },
                "parent": {
                    "type": "string"
                },
                "previous": {
                    "type": "string"
                },
                "sat": {
                    "type": "integer"
                },
                "satPoint": {
                    "type": "string"
                },
                "timestamp": {
                    "type": "integer"
                },
                "value": {
                    "type": "integer"
                }
            }
        }
    }
}